requirements.txt
dbt-materialize==1.0.4
streamlit
kafka-python
websocket-client